refactor: reorganize modal components and update imports for consistency
