+++ src/features/edit/components/index.ts
@@ -1,8 +1,8 @@
 export { EditPreview } from './EditPreview';
-export { ViewMarcModal } from './ViewMarcModal';
 export { EditSection } from './EditSection';
 export type { IDrawComponent } from './EditSection';
 export { EditControlPane } from './EditControlPane';
 export { RecordControls } from './RecordControls';
 export { ViewMarcControlPane } from './ViewMarcControlPane';
 export { InstancesList } from './InstancesList';
+export * from './modals';
