+++ src/features/edit/components/modals/ModalCloseRecord/ModalCloseRecord.scss
