+++ src/features/edit/components/modals/ModalCloseRecord/ModalCloseRecord.tsx
