+++ src/features/edit/components/modals/ModalSwitchToNewRecord/ModalSwitchToNewRecord.scss
