+++ src/features/edit/components/modals/ModalViewMarc/index.ts
@@ -1,0 +1,1 @@
+export { ModalViewMarc } from './ModalViewMarc';
