+++ src/features/edit/components/modals/Prompt/Prompt.scss
