+++ src/features/edit/components/modals/index.ts
@@ -1,0 +1,6 @@
+export { ModalCloseRecord } from './ModalCloseRecord';
+export { ModalDeleteRecord } from './ModalDeleteRecord';
+export { ModalSwitchToNewRecord } from './ModalSwitchToNewRecord';
+export { ModalUncontrolledAuthorities } from './ModalUncontrolledAuthorities';
+export { ModalViewMarc } from './ModalViewMarc';
+export { Prompt } from './Prompt';
